requirements.txt
flask
mysql-connector-python
flask-bcrypt
